Update types.tex
Added to Tuples example,lines 306-309

Now shows a tuple being created, used as a parameter, deconstructed and returned as a new tuple. Recommend keeping the comment as the phrase does not come up anywhere in the doc. I had to google it, my google foo was good...

This editing change was pretty easy to do, I'll send more your way as I find them.