Update types.tex
Fixed four typos, calling int "uint", added colon after the function signature and removed the "es" within the type Graph declaration.

Added the main method to show the setup to use the depthFirstSearch and better illustrate the two sets, how they are created and used. It took me an hour to figure it out... This may not fit the style of document you want, I've put it in so you have the choice. As it is, you can copy & paste it into Whiley play.