Updated GettingStartedWithWhiley
This updates the "Getting Started" tutorial in several ways.  Firstly, a
build script is introduced to ensure that all examples are now compiled
using the latest version of the Whiley compiler.  Secondly, I've updated
many of the examples and various bits of text to reflect the latest
version of the language syntax.